+++ src/main/java/com/watchmyai/quota/UsageSnapshot.java
@@ -1,9 +1,11 @@
 package com.watchmyai.quota;
 
+import java.math.BigDecimal;
+
 public record UsageSnapshot(
         int usedLifetimeRequests,
         int usedMonthlyRequests,
         int usedPremiumRequests,
-        double estimatedMonthlyCostEur
+        BigDecimal estimatedMonthlyCostEur
 ) {
 }
